feat: add compile and test stages (#7) (#20)
The pipeline clones the git repo and runs:
- mvn compile
- mvn test

Commit status is updated accordingly and build logs are stored in the
database when the pipeline finishes with or without a failure.

The unit tests for the pipeline is omitted in this commit since the
team has to continue working with these changes in main branch. But,
the ci server has been regression tested on the GCP Linux server it
will run on later after these changes has been added. It works what we
can see at least.

Unit tests will be added in a later commit.